+++ speeduino/speeduino.h
@@ -22,7 +22,7 @@ byte getVE1(void);
 byte getAdvance1(void);
 uint16_t calculatePWLimit();
 void calculateStaging(uint32_t);
-void calculateIgnitionAngles(int dwellAngle);
+void calculateIgnitionAngles(uint16_t dwellAngle);
 void checkLaunchAndFlatShift();
 
 extern uint16_t req_fuel_uS; /**< The required fuel variable (As calculated by TunerStudio) in uS */
